backend-python/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0
apscheduler==3.10.4
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
pykrx==1.0.45
finance-datareader==0.9.60
plotly==5.18.0
